win32: proper libc-less version.
While the portable interface is KERNEL32.dll (real syscalls are undocumented
and break compat even on minor releases of Windows), there are many libcs;
msvcrt.dll has kept compat nicely but it's still safer to not rely on it.

As a bonus, a C compiler is no longer needed, merely binutils.